# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DomusUnify API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/calendar/events": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "summary": "Lista eventos do calendário visíveis para o utilizador autenticado.",
        "description": "Pode filtrar por:\r\n- dateUtc (dia inteiro, UTC)\r\n- intervalo (fromUtc/toUtc)\r\n- search (parte do título)\r\n- participantUserId (participante específico)\r\n- take (limitar o número de eventos devolvidos)\r\n            \r\nSe nenhum filtro for fornecido, devolve uma janela default (ex: últimos 7 dias + próximos 30).",
        "parameters": [
          {
            "name": "fromUtc",
            "in": "query",
            "description": "Data/hora inicial (UTC) para filtrar por intervalo, opcional.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toUtc",
            "in": "query",
            "description": "Data/hora final (UTC) para filtrar por intervalo, opcional.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateUtc",
            "in": "query",
            "description": "Dia (UTC) para filtrar por dia inteiro, opcional.",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2026-02-04"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Texto para procurar no título, opcional.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "participantUserId",
            "in": "query",
            "description": "Filtra por participante específico, opcional.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "take",
            "in": "query",
            "description": "Número máximo de eventos a devolver. Se não for fornecido, devolve todos os eventos que correspondem aos outros filtros.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEventResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEventResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEventResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Calendar"
        ],
        "summary": "Cria um novo evento do calendário para a família atual.",
        "description": "Permite definir participantes, visibilidade, lembretes, recorrência, etc. O utilizador deve ter permissões na família.",
        "requestBody": {
          "description": "O pedido contendo detalhes do evento como título, datas, participantes, etc.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calendar/events/{eventId}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "summary": "Devolve os detalhes de um evento do calendário pelo seu ID.",
        "description": "Retorna `404 Not Found` se o evento não existir ou não for visível para o utilizador autenticado na família atual.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "O ID do evento.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventDetailResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventDetailResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventDetailResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Calendar"
        ],
        "summary": "Atualiza um evento do calendário existente.",
        "description": "Permite alterar título, datas, participantes, etc. O utilizador deve ter permissões para editar o evento.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "O ID do evento a atualizar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes atualizados do evento.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendar"
        ],
        "summary": "Elimina um evento do calendário.",
        "description": "O utilizador deve ter permissões para editar o evento na família atual.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "O ID do evento a eliminar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/calendar/events/{eventId}/duplicate": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "summary": "Duplica um evento do calendário com novas horas de início e fim.",
        "description": "Cria uma cópia do evento com datas ajustadas. O utilizador deve ter permissões na família.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "O ID do evento a duplicar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo novas horas de início e fim.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateCalendarEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateCalendarEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateCalendarEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/calendar/events/{eventId}/recurrence": {
      "patch": {
        "tags": [
          "Calendar"
        ],
        "summary": "Atualiza um evento recorrente do calendário.",
        "description": "Permite editar a série recorrente ou uma ocorrência específica. O scope define se é toda a série ou apenas uma ocorrência.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "O ID do evento recorrente.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes da atualização recorrente.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecurringEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecurringEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecurringEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DuplicateCalendarEventRequest": {
        "type": "object",
        "properties": {
          "newStartUtc": {
            "type": "string",
            "description": "Novo início (UTC) para o evento duplicado.",
            "format": "date-time"
          },
          "newEndUtc": {
            "type": "string",
            "description": "Novo fim (UTC) para o evento duplicado.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Pedido para duplicar um evento numa nova data/hora."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Insere o token JWT assim: **Bearer {token}**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}